@@ -1,4 +1,4 @@
-﻿using Discord.WebSocket;
+﻿using NetCord.Services.ApplicationCommands;
 using Melpominee.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -12,12 +12,8 @@ namespace Melpominee
             using IHost host = Host.CreateDefaultBuilder(args)
             .ConfigureServices(services =>
             {
-                services.AddSingleton<AudioService>();
-                services.AddSingleton<DiscordSocketClient>();
-                services.AddSingleton((p) => DataContext.Instance);
-                services.AddHostedService((p) => p.GetRequiredService<AudioService>());
-                services.AddHostedService<CommandHandler>();
-                services.AddHostedService<InteractionHandler>();
+                //services.AddSingleton((p) => DataContext.Instance);
+                services.AddSingleton<MelpomineeAudioService>();
                 services.AddHostedService<MelpomineeService>();
             })
             .Build();
